"Bundle" the extension, and migrate to ESLint
Bundling the extension makes it more efficient by combining all the JS files (of which I have 1,600), reducing the file count and thus network requests when running in a web or electron context. And I think this is enabling minification too.

Migrates from TSLint to ESLint, because TSLint is deprecated.