symbols: more concise display of signature as details
Removes the redundant display of the name and shows only the argins and argouts as the details. Replaces "varargin" with "..." and "this" with "_" to be even tighter.

Would probably want to make this configurable at some point, so you can turn the varargin and this replacements on or off, define what conventional dispatch object name you use (e.g. "this", "obj", "self", etc.), and maybe support custom replacement strings.